Delete old saved queries before downloading from BigQuery; README updated.
Now the scripts remove old versions of views and scheduled queries if a new version
is requested. README includes information about this and now explains the keys in
the config file.

NOTE: as before, the create_flowchart script updates views and scheduled queries if
requested. So it is a full backup of BigQuery.